Fix NoClassDefFoundError on Android < 8.0
